무비 디테일에 새로운 훅 만들어서 작성, 피드백 반영 완료
